Lobby + Game Enviroment
I have just finished setting the lobby, though it still doesn't have the XR versions in it. All the 3D related and all the modeling related stuff are done. There are some more photos to be added, but it's not a big deal. Also I added the environment for the game. I added it under the GameObject "Environment", if you hide it everything will be hidden. I didn't change anything in the game settings, I just added the visual stuff. Plus I added a bit of the lighting because there was no lighting.